Update monopoly.html
Added features and the board prototype is almost finished. Added a lot more styling, and a <script> section which will later be put into a different file. <style> section most likely will too just to minimize content of the file. On CodeAcademy, I'm starting to learn JavaScript. A lot of the syntax I use is from W3Schools.