Update monopoly.js
added doubles and being put in jail after getting 3 doubles. There is currently no way to get out of jail, and the third person to roll a double is put in jail (this is for testing purposes as it's almost impossible to roll 3 doubles in a row. This must be fixed for the final version, and returning the variable doublesRolled to 0 after each player's turn must also be added).